python_tests/code/requirements.txt
attrs==21.2.0
pytest-xdist==2.4.0
requests==2.22.0
selenium==3.14.1
webdriver-manager==3.3.0
pymysql==1.0.2
SQLAlchemy==1.4.35
faker==13.3.1
werkzeug==2.0.3
psycopg2-binary==2.9.3